+++ vss_cli/plugins/compute_plugins/folder.py
@@ -7,7 +7,7 @@ import vss_cli.autocompletion as autocompletion
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
 from vss_cli.exceptions import VssCliError
-from vss_cli.helper import format_output, process_filters
+from vss_cli.helper import format_output
 from vss_cli.plugins.compute import cli
 
 _LOGGING = logging.getLogger(__name__)
@@ -36,15 +36,15 @@ def compute_folder_ls(ctx: Configuration, filter_by, show_all, sort, page):
 
             vss-cli compute folder ls -f name=like,%Project%
 
-        Sort list in the following format <field_name> <asc|desc>. For example:
+        Sort list in the following format <field_name>=<asc|desc>. For example:
 
-            vss-cli compute folder ls -s path desc
+            vss-cli compute folder ls -s path=desc
     """
     params = dict(expand=1, sort='path,asc')
     if all(filter_by):
-        params['filter'] = ';'.join(process_filters(filter_by))
+        params['filter'] = ';'.join(filter_by)
     if all(sort):
-        params['sort'] = f'{sort[0]},{sort[1]}'
+        params['sort'] = ';'.join(sort)
     # get objects
     with ctx.spinner(disable=ctx.debug):
         obj = ctx.get_folders(show_all=show_all, **params)