+++ vss_cli/plugins/compute_plugins/image.py
@@ -5,7 +5,7 @@ import click
 from vss_cli import const, rel_opts as so
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
-from vss_cli.helper import format_output, process_filters
+from vss_cli.helper import format_output
 from vss_cli.plugins.compute import cli
 
 _LOGGING = logging.getLogger(__name__)
@@ -37,13 +37,13 @@ def compute_image_public_ls(
 
     Filter by name and sort desc. For example:
 
-        vss-cli compute image public ls -f name=like,Cent% -s path asc
+        vss-cli compute image public ls -f name=like,Cent% -s path=asc
     """
     params = dict(expand=1, sort='name,asc')
     if all(filter_by):
-        params['filter'] = ';'.join(process_filters(filter_by))
+        params['filter'] = ';'.join(filter_by)
     if all(sort):
-        params['sort'] = f'{sort[0]},{sort[1]}'
+        params['sort'] = ';'.join(sort)
     # get objects
     with ctx.spinner(disable=ctx.debug):
         obj = ctx.get_images(show_all=show_all, **params)