+++ vss_cli/plugins/compute_plugins/template.py
@@ -5,7 +5,7 @@ import click
 from vss_cli import const, rel_opts as so
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
-from vss_cli.helper import format_output, process_filters
+from vss_cli.helper import format_output
 from vss_cli.plugins.compute import cli
 
 _LOGGING = logging.getLogger(__name__)
@@ -36,14 +36,14 @@ def compute_template_ls(
 
         Simple name filtering:
 
-        vss-cli compute template ls -f name=%vm-name% -s name desc
+        vss-cli compute template ls -f name=%vm-name% -s name=desc
 
     """
     params = dict(expand=1, sort='name,asc')
     if all(filter_by):
-        params['filter'] = ';'.join(process_filters(filter_by))
+        params['filter'] = ';'.join(filter_by)
     if all(sort):
-        params['sort'] = f'{sort[0]},{sort[1]}'
+        params['sort'] = ';'.join(sort)
     # get templates
     with ctx.spinner(disable=ctx.debug):
         obj = ctx.get_templates(show_all=show_all, per_page=count, **params)