+++ vss_cli/plugins/message.py
@@ -5,7 +5,7 @@ from vss_cli import const, rel_opts as so
 import vss_cli.autocompletion as autocompletion
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
-from vss_cli.helper import format_output, process_filters
+from vss_cli.helper import format_output
 
 
 @click.group('message', short_help='Manage VSS Messages.')
@@ -32,17 +32,17 @@ def message_ls(ctx: Configuration, filter_by, page, sort, show_all, count):
 
             vss-cli message ls -f kind=eq,Notice
 
-        Sort list in the following format <field_name> <asc|desc>. For example:
+        Sort list in the following format <field_name>=<asc|desc>. For example:
 
-            vss-cli message ls -s created_on desc
+            vss-cli message ls -s created_on=desc
 
     """
     columns = ctx.columns or const.COLUMNS_MESSAGE_MIN
     params = dict(expand=1, sort='status,asc')
     if all(filter_by):
-        params['filter'] = ';'.join(process_filters(filter_by))
+        params['filter'] = ';'.join(filter_by)
     if all(sort):
-        params['sort'] = f'{sort[0]},{sort[1]}'
+        params['sort'] = ';'.join(sort)
     # make request
     with ctx.spinner(disable=ctx.debug):
         _requests = ctx.get_user_messages(