+++ vss_cli/plugins/request_plugins/new.py
@@ -33,22 +33,22 @@ def request_mgmt_new_ls(
 
         Filter list in the following format <field_name> <operator>,<value>
         where operator is eq, ne, lt, le, gt, ge, like, in.
-        For example: status eq,PROCESSED
+        For example: status=eq,PROCESSED
 
-            vss-cli request new ls -f status eq,PROCESSED
+            vss-cli request new ls -f status=eq,PROCESSED
 
-        Sort list in the following format <field_name> <asc|desc>. For example:
+        Sort list in the following format <field_name>=<asc|desc>. For example:
 
-            vss-cli request new ls -s created_on desc
+            vss-cli request new ls -s created_on=desc
 
     """
     columns = ctx.columns or const.COLUMNS_REQUEST_NEW_MIN
     _LOGGING.debug(f'Columns {columns}')
     params = dict(expand=1, sort='created_on,desc')
     if all(filter_by):
-        params['filter'] = f'{filter_by[0]},{filter_by[1]}'
+        params['filter'] = ';'.join(filter_by)
     if all(sort):
-        params['sort'] = f'{sort[0]},{sort[1]}'
+        params['sort'] = ';'.join(sort)
     # make request
     with ctx.spinner(disable=ctx.debug):
         _requests = ctx.get_new_requests(